requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cmake==3.20.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.10.0
defusedxml==0.7.1
Django==3.0.3
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
djoser==2.1.0
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
imutils==0.5.4
itypes==1.2.0
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.20.3
oauthlib==3.1.0
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.52
pandas==1.2.4
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
pynotify==0.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
uWSGI==2.0.19.1
